split code for version, follow ups and changes
