+++ hat/assets/js/apps/Iaso/domains/workflows/hooks/requests/useUpdateWorkflowChange.tsx
@@ -2,7 +2,7 @@ import { UseMutationResult } from 'react-query';
 import { postRequest } from '../../../../libs/Api';
 import { useSnackMutation } from '../../../../libs/apiHooks';
 
-import { Change } from '../../types/workflows';
+import { Change } from '../../types';
 
 const updateWorkflowChange = async (data: Change): Promise<any> =>
     postRequest(`/api/workflowChanges/${data.id}/`, data);
